app/undo_history.c app/core/gimpbrush.c app/core/gimpimage-new.c
2003-01-05  Manish Singh  <yosh@gimp.org>

        * app/undo_history.c
        * app/core/gimpbrush.c
        * app/core/gimpimage-new.c
        * app/core/gimpobject.c
        * app/core/gimppalette-import.c
        * app/core/gimppattern.c
        * app/plug-in/plug-in.c
        * app/tools/gimpbezierselecttool.c
        * libgimpwidgets/gimpunitmenu.c
        * plug-ins/MapObject/mapobject_ui.c
        * plug-ins/common/convmatrix.c
        * plug-ins/common/curve_bend.c
        * plug-ins/common/sample_colorize.c
        * plug-ins/common/tiff.c
        * plug-ins/flame/flame.c
        * plug-ins/gflare/gflare.c
        * plug-ins/gimpressionist/general.c
        * plug-ins/gimpressionist/orientation.c
        * plug-ins/gimpressionist/preview.c
        * plug-ins/gimpressionist/size.c
        * plug-ins/imagemap/imap_grid.c
        * plug-ins/imagemap/imap_menu.c
        * plug-ins/maze/algorithms.c
        * plug-ins/script-fu/interp_regex.c
        * plug-ins/script-fu/interp_sliba.c
        * plug-ins/script-fu/script-fu-console.c
        * plug-ins/script-fu/script-fu-server.c
        * plug-ins/webbrowser/webbrowser.c: added GINT_TO_POINTER and friends,
        fixed format strings, for 64-bitness.

        * modules/colorsel_triangle.c
        * plug-ins/tools/tool-safe-mode-plug-in.c: #include missing header
        files

+++ app/core/gimpbrush-load.c
@@ -512,7 +512,7 @@ gimp_brush_load_brush (gint          fd,
     {
       g_set_error (error, GIMP_DATA_ERROR, GIMP_DATA_ERROR_READ,
                    _("Could not read %d bytes from '%s': %s"),
-                   sizeof (header), filename, g_strerror (errno));
+                   (gint) sizeof (header), filename, g_strerror (errno));
       return NULL;
     }
