+++ app/core/gimppattern.c
@@ -309,7 +309,7 @@ gimp_pattern_load (const gchar  *filename,
     {
       g_set_error (error, GIMP_DATA_ERROR, GIMP_DATA_ERROR_READ,
                    _("Could not read %d bytes from '%s': %s"),
-                   sizeof (header), filename, g_strerror (errno));
+                   (gint) sizeof (header), filename, g_strerror (errno));
       goto error;
     }
