+++ plug-ins/common/tiff.c
@@ -698,13 +698,13 @@ load_rgba (TIFF *tif, channel_data *channel)
 {
   uint32 imageWidth, imageLength;
   uint32 row;
-  gulong *buffer;
+  uint32 *buffer;
 
   TIFFGetField(tif, TIFFTAG_IMAGEWIDTH, &imageWidth);
   TIFFGetField(tif, TIFFTAG_IMAGELENGTH, &imageLength);
   gimp_pixel_rgn_init (&(channel[0].pixel_rgn), channel[0].drawable,
                           0, 0, imageWidth, imageLength, TRUE, FALSE);
-  buffer = g_new(gulong, imageWidth * imageLength);
+  buffer = g_new(uint32, imageWidth * imageLength);
   channel[0].pixels = (guchar*) buffer;
   if (buffer == NULL) {
     g_message("TIFF Unable to allocate temporary buffer\n");
